dns/asyncio/udp: make the ring's new surfaces work on Windows
The four `tests/nimony/async` tests ran their real bodies on POSIX and
printed their goldens verbatim on Windows. They run for real on both
backends now (IOCP and `-d:nimIoringWsaPoll`), so the stubs are gone.

hexer, and the reason none of this could work:

- `eraiser.trAsgn` wrapped a raising routine's `result = v` in
  `(tupconstr (tuple ErrorCode T) Success v)`. `cps` decides where one
  coroutine state ends by looking at the ROOT of a statement, so a
  `.passive` call one level down inside that constructor was invisible:
  the state boundary landed after the whole assignment, which then sat on
  the far side of the transition and never ran. `result = recvFrom(...)`
  returned zero, always — and where the call was in tail position it did
  not compile at all (`cps.nim: state != -1`). The two halves are assigned
  separately now, which says the same thing, builds no temporary tuple,
  and leaves the call where `cps` looks for it.
- `coro_transform.trGoto` rejects the nested shape outright rather than
  emitting an assignment it cannot place.

std/ioring, Windows:

- The IOCP drain handed `opSocket` — whose fd is `-1`, its socket not yet
  created — to `ensureAssociated`, which failed and dropped the op without
  completing it: every passive socket creation parked until its deadline.
  The dispatch now names the kinds that must not be associated, and a
  failed association completes `ECancelled` again instead of leaking the
  slot (a regression against master's arm).
- `opConnect` always went through ConnectEx, which is defined for
  connection-oriented sockets and fails a SOCK_DGRAM with WSAEINVAL. A
  datagram connect records a peer and answers at once, so it is the plain
  Winsock `connect`, completed in place.
- `GetFileType` ran on every socket op; only a read or a write can be on a
  file handle, so only those ask.
- backends/files.nim narrowed a pointer-sized HANDLE through `uint32`,
  which warns on every use.

std/socket: `addrFromHost` and `wildcardAddr` wrote the BSD `sa_len`
layout on Windows, where `sin_family` is a 2-byte field at offset 0 — so
every address they built named family 528. The family bytes live in one
place now.

std/asyncio: `win32Mode` returned a tuple over winlean's `DWORD`, whose
`importc` pragmas end up in the C type's name, so the two spellings of
that tuple would not assign to each other.

std/dns: the resolver read `/etc/hosts` and `/etc/resolv.conf`, neither of
which exists on Windows. The hosts file is read from `%SystemRoot%` there,
the nameservers come from `GetNetworkParams`, and `localhost` resolves to
loopback by rule (RFC 6761) rather than by table entry — stock Windows
ships both of its `localhost` lines commented out. `Resolver.hosts` is
always constructed and `DnsRequest.sock` is `nil ptr`, which is what the
merge with master's not-nil pointers asks for.

+++ lib/std/ioring/backends/files.nim
@@ -22,7 +22,7 @@
 # surfaces make: kernel handle values are small 4-aligned integers in practice
 # (undocumented), and a HANDLE too large to narrow fails the open rather than
 # being abused. `GetFileType` is the reliability the narrowing relies on: it
-# returns FILE_TYPE_DISK for a regular file and FILE_TYPE_UNKNOWN for a socket,
+# returns FILE_TYPE_DISK for a regular file and something else for a socket,
 # which is how a read/write op tells a file Handle from a SOCKET in the shared
 # cint space — a registry-free answer that stays correct when a cint value is
 # reused by the other kind of descriptor.
@@ -44,14 +44,15 @@ when defined(windows):
   proc handleOf(fd: cint): Handle {.inline.} =
     ## Widen the ring's cint narrowing back to a HANDLE without sign
     ## extension (a file entered the cint space by the same narrowing).
-    cast[Handle](cast[uint32](fd))
+    cast[Handle](uint(cast[uint32](fd)))   # via `uint`: a HANDLE is pointer-sized
 
   proc isFileHandle*(fd: cint): bool {.inline.} =
     ## True when `fd` names a Win32 file HANDLE rather than a Winsock SOCKET.
-    ## GetFileType answers: a regular file is FILE_TYPE_DISK, and a socket —
-    ## which is not a file — reports FILE_TYPE_UNKNOWN. Deliberately no wider
-    ## chart: a pipe or console would be served fine by the sync path, but the
-    ## ring's non-socket descriptors are the files asyncio opens.
+    ## GetFileType answers: a regular file is FILE_TYPE_DISK and a socket is
+    ## not (it is an AFD file object, and reports FILE_TYPE_PIPE). Deliberately
+    ## no wider chart: matching DISK exactly is what makes every other answer —
+    ## pipe, char device, unknown — mean "not one of ours", so the test cannot
+    ## be wrong about a descriptor kind it was not told about.
     getFileType(handleOf(fd)) == FILE_TYPE_DISK
 
   proc completeFileOpen*(idx: int; path: cstring; desiredAccess, disposition: int32) =
@@ -72,7 +73,7 @@ when defined(windows):
       discard closeHandle(h)   # cannot be narrowed to the ring's cint fd space
       complete(idx, -1)
     else:
-      complete(idx, int(cast[uint32](h)))
+      complete(idx, int(uint32(cast[uint](h))))
 
   proc completeFileRead*(idx: int; fd: cint; buf: pointer; len: int) =
     ## One synchronous `ReadFile` on the polling thread, completing the op
